feat: add crypto-context.jsx
