+++ src/App.jsx
@@ -3,18 +3,21 @@ import AppContent from './components/layout/AppContent';
 import AppFooter from './components/layout/AppFooter';
 import AppHeader from './components/layout/AppHeader';
 import AppSidebar from './components/layout/AppSidebar';
+import { CryptoContextProvider } from './context/crypto-context';
 
 export default function App() {
   return (
     <>
-      <Layout>
-        <AppHeader />
+      <CryptoContextProvider>
         <Layout>
-          <AppSidebar />
-          <AppContent />
+          <AppHeader />
+          <Layout>
+            <AppSidebar />
+            <AppContent />
+          </Layout>
+          <AppFooter />
         </Layout>
-        <AppFooter />
-      </Layout>
+      </CryptoContextProvider>
     </>
   );
 }
