+++ src/context/crypto-context.jsx
@@ -1,0 +1,47 @@
+import { createContext, useEffect, useState } from 'react';
+import { fetchAssets, fetchCrypto } from '../components/layout/api';
+import { percentDifference } from '../components/utils';
+
+const CryptoContext = createContext({
+  assets: [],
+  crypto: [],
+  loading: false,
+});
+
+export function CryptoContextProvider({ children }) {
+  const [loading, setLoading] = useState(false);
+  const [crypto, setCrypto] = useState([]);
+  const [assets, setAssets] = useState([]);
+
+  useEffect(() => {
+    async function preload() {
+      setLoading(true);
+      const { result } = await fetchCrypto();
+      const assets = await fetchAssets();
+
+      setAssets(
+        assets.map(asset => {
+          const coin = result.find(c => c.id === asset.id);
+          return {
+            grow: asset.price < coin.price,
+            growPercent: percentDifference(asset.price, coin.price),
+            totalAmount: asset.amount * coin.price,
+            totalProfit: asset.amount * coin.price - asset.amount * asset.price,
+            ...asset,
+          };
+        })
+      );
+      setCrypto(result);
+      setLoading(false);
+    }
+    preload();
+  }, []);
+
+  return (
+    <CryptoContext.Provider value={{ loading, crypto, assets }}>
+      {children}
+    </CryptoContext.Provider>
+  );
+}
+
+export default CryptoContext;
